fix review issues: invulnerability check, image caching, missing fields, stale test
- add Invulnerable check in KillPlayer
- cache ability images in Renderer constructor
- restore Username/Disconnected in PlayerDto
- delete unused AbilityState.cs and stale PlayerStateBoostTest
- add EOF newlines

+++ backend/Game/GameRoom.cs
@@ -301,7 +301,7 @@ public class GameRoom
                 Alive = ps.IsAlive,
                 Disconnected = ps.IsDisconnected,
                 ColorId = ps.ColorId,
-                SpeedMultiplier = ps.SpeedMultiplier,
+                SpeedMultiplier = (int)ps.SpeedMultiplier,
                 Abilities = playerAbilities
             });
         }
